Upgrade to RN 39
