+++ android/settings.gradle
@@ -1,6 +1,5 @@
 rootProject.name = 'ReactNativeFSTest'
 
 include ':app'
-
 include ':react-native-fs'
-project(':react-native-fs').projectDir = new File(settingsDir, '../node_modules/react-native-fs/android')
+project(':react-native-fs').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-fs/android')
