+++ package.json
@@ -7,8 +7,8 @@
     "test": "mocha test/node-cli.js"
   },
   "dependencies": {
-    "react": "15.3.1",
-    "react-native": "0.36.1",
+    "react": "15.4.1",
+    "react-native": "0.39.0",
     "react-native-fs": "git+https://github.com/johanneslumpe/react-native-fs.git"
   },
   "devDependencies": {
